+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # CHANGELOG
 
-## 14/04/2017 [Dani 6h]
+## 14/04/2017 [Dani 8h]
 
 * Lik03
 * Rename webapp to app-visuals and refactor app
+* Add keyboard to control-app
+* Add group mode to control-app
 
 ## 12/04/2017